Reorganize 'funcx_sdk' tests so that tox is better
- move tests out of the `funcx` directory tree so that we aren't
  unnecessarily distributing tests as part of the packaged application
- move any tests which do not run under `tox` into `tests/archived/`
- add `tests/archived/` to pytest `norecursedirs`, meaning "pytest,
  don't look here"
- simplify the `tox` invocation of `pytest` to drop specific test
  directory names, allowing `tox -- tests/test_foo.py` to run a
  specific module
