+++ funcx_sdk/setup.cfg
@@ -13,3 +13,6 @@ ignore_missing_imports = true
 warn_unreachable = true
 warn_no_return = true
 exclude = tests
+
+[tool:pytest]
+norecursedirs = tests/archived
