+++ funcx_sdk/tests/archived/test_errors/test_invalid_endpoint.py
