+++ funcx_sdk/tests/unit/test_client_init.py
