+++ funcx_sdk/tests/unit/test_environment_lookups.py
