+++ funcx_sdk/tox.ini
@@ -5,7 +5,7 @@ skip_missing_interpreters = true
 [testenv]
 usedevelop = true
 extras = test
-commands = pytest funcx/tests/unit/ funcx/tests/integration/ {posargs}
+commands = pytest {posargs}
 
 [testenv:mypy]
 deps = mypy
